+++ azurerm/resource_arm_automation_credential.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/Azure/azure-sdk-for-go/arm/automation"
+	"github.com/Azure/azure-sdk-for-go/services/automation/mgmt/2015-10-31/automation"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
@@ -52,10 +52,12 @@ func resourceArmAutomationCredential() *schema.Resource {
 
 func resourceArmAutomationCredentialCreateUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*ArmClient).automationCredentialClient
+	ctx := meta.(*ArmClient).StopContext
 	log.Printf("[INFO] preparing arguments for AzureRM Automation Credential creation.")
 
 	name := d.Get("name").(string)
 	resGroup := d.Get("resource_group_name").(string)
+	client.ResourceGroupName = resGroup
 	accName := d.Get("account_name").(string)
 	user := d.Get("username").(string)
 	password := d.Get("password").(string)
@@ -70,12 +72,12 @@ func resourceArmAutomationCredentialCreateUpdate(d *schema.ResourceData, meta in
 		Name: &name,
 	}
 
-	_, err := client.CreateOrUpdate(resGroup, accName, name, parameters)
+	_, err := client.CreateOrUpdate(ctx, accName, name, parameters)
 	if err != nil {
 		return err
 	}
 
-	read, err := client.Get(resGroup, accName, name)
+	read, err := client.Get(ctx, accName, name)
 	if err != nil {
 		return err
 	}
@@ -91,16 +93,17 @@ func resourceArmAutomationCredentialCreateUpdate(d *schema.ResourceData, meta in
 
 func resourceArmAutomationCredentialRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*ArmClient).automationCredentialClient
+	ctx := meta.(*ArmClient).StopContext
 	id, err := parseAzureResourceID(d.Id())
 	if err != nil {
 		return err
 	}
 	resGroup := id.ResourceGroup
+	client.ResourceGroupName = resGroup
 	accName := id.Path["automationAccounts"]
 	name := id.Path["credentials"]
 
-	resp, err := client.Get(resGroup, accName, name)
-
+	resp, err := client.Get(ctx, accName, name)
 	if err != nil {
 		if utils.ResponseWasNotFound(resp.Response) {
 			d.SetId("")
@@ -123,17 +126,18 @@ func resourceArmAutomationCredentialRead(d *schema.ResourceData, meta interface{
 
 func resourceArmAutomationCredentialDelete(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*ArmClient).automationCredentialClient
+	ctx := meta.(*ArmClient).StopContext
 
 	id, err := parseAzureResourceID(d.Id())
 	if err != nil {
 		return err
 	}
 	resGroup := id.ResourceGroup
+	client.ResourceGroupName = resGroup
 	accName := id.Path["automationAccounts"]
 	name := id.Path["credentials"]
 
-	resp, err := client.Delete(resGroup, accName, name)
-
+	resp, err := client.Delete(ctx, accName, name)
 	if err != nil {
 		if utils.ResponseWasNotFound(resp) {
 			return nil