+++ azurerm/resource_arm_image.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/Azure/azure-sdk-for-go/arm/compute"
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2017-03-30/compute"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
@@ -147,8 +147,8 @@ func resourceArmImage() *schema.Resource {
 }
 
 func resourceArmImageCreateUpdate(d *schema.ResourceData, meta interface{}) error {
-	client := meta.(*ArmClient)
-	imageClient := client.imageClient
+	client := meta.(*ArmClient).imageClient
+	ctx := meta.(*ArmClient).StopContext
 
 	log.Printf("[INFO] preparing arguments for AzureRM Image creation.")
 
@@ -206,13 +206,17 @@ func resourceArmImageCreateUpdate(d *schema.ResourceData, meta interface{}) erro
 		ImageProperties: &properties,
 	}
 
-	_, imageErr := imageClient.CreateOrUpdate(resGroup, name, createImage, make(chan struct{}))
-	err = <-imageErr
+	future, err := client.CreateOrUpdate(ctx, resGroup, name, createImage)
 	if err != nil {
 		return err
 	}
 
-	read, err := imageClient.Get(resGroup, name, "")
+	err = future.WaitForCompletion(ctx, client.Client)
+	if err != nil {
+		return err
+	}
+
+	read, err := client.Get(ctx, resGroup, name, "")
 	if err != nil {
 		return err
 	}
@@ -226,7 +230,8 @@ func resourceArmImageCreateUpdate(d *schema.ResourceData, meta interface{}) erro
 }
 
 func resourceArmImageRead(d *schema.ResourceData, meta interface{}) error {
-	imageClient := meta.(*ArmClient).imageClient
+	client := meta.(*ArmClient).imageClient
+	ctx := meta.(*ArmClient).StopContext
 
 	id, err := parseAzureResourceID(d.Id())
 	if err != nil {
@@ -235,7 +240,7 @@ func resourceArmImageRead(d *schema.ResourceData, meta interface{}) error {
 	resGroup := id.ResourceGroup
 	name := id.Path["images"]
 
-	resp, err := imageClient.Get(resGroup, name, "")
+	resp, err := client.Get(ctx, resGroup, name, "")
 	if err != nil {
 		if utils.ResponseWasNotFound(resp.Response) {
 			d.SetId("")
@@ -246,20 +251,23 @@ func resourceArmImageRead(d *schema.ResourceData, meta interface{}) error {
 
 	d.Set("name", resp.Name)
 	d.Set("resource_group_name", resGroup)
-	d.Set("location", azureRMNormalizeLocation(*resp.Location))
+
+	if location := resp.Location; location != nil {
+		d.Set("location", azureRMNormalizeLocation(*location))
+	}
 
 	//either source VM or storage profile can be specified, but not both
 	if resp.SourceVirtualMachine != nil {
 		d.Set("source_virtual_machine_id", resp.SourceVirtualMachine.ID)
 	} else if resp.StorageProfile != nil {
 		if disk := resp.StorageProfile.OsDisk; disk != nil {
-			if err := d.Set("os_disk", flattenAzureRmImageOSDisk(d, disk)); err != nil {
+			if err := d.Set("os_disk", flattenAzureRmImageOSDisk(disk)); err != nil {
 				return fmt.Errorf("[DEBUG] Error setting AzureRM Image OS Disk error: %+v", err)
 			}
 		}
 
 		if disks := resp.StorageProfile.DataDisks; disks != nil {
-			if err := d.Set("data_disk", flattenAzureRmImageDataDisks(d, disks)); err != nil {
+			if err := d.Set("data_disk", flattenAzureRmImageDataDisks(disks)); err != nil {
 				return fmt.Errorf("[DEBUG] Error setting AzureRM Image Data Disks error: %+v", err)
 			}
 		}
@@ -271,7 +279,8 @@ func resourceArmImageRead(d *schema.ResourceData, meta interface{}) error {
 }
 
 func resourceArmImageDelete(d *schema.ResourceData, meta interface{}) error {
-	imageClient := meta.(*ArmClient).imageClient
+	client := meta.(*ArmClient).imageClient
+	ctx := meta.(*ArmClient).StopContext
 
 	id, err := parseAzureResourceID(d.Id())
 	if err != nil {
@@ -280,8 +289,12 @@ func resourceArmImageDelete(d *schema.ResourceData, meta interface{}) error {
 	resGroup := id.ResourceGroup
 	name := id.Path["images"]
 
-	_, deleteErr := imageClient.Delete(resGroup, name, make(chan struct{}))
-	err = <-deleteErr
+	future, err := client.Delete(ctx, resGroup, name)
+	if err != nil {
+		return err
+	}
+
+	err = future.WaitForCompletion(ctx, client.Client)
 	if err != nil {
 		return err
 	}
@@ -289,18 +302,18 @@ func resourceArmImageDelete(d *schema.ResourceData, meta interface{}) error {
 	return nil
 }
 
-func flattenAzureRmImageOSDisk(d *schema.ResourceData, osDisk *compute.ImageOSDisk) []interface{} {
+func flattenAzureRmImageOSDisk(osDisk *compute.ImageOSDisk) []interface{} {
 	result := make(map[string]interface{})
 
 	if disk := osDisk; disk != nil {
-		if osDisk.BlobURI != nil {
-			result["blob_uri"] = *osDisk.BlobURI
+		if uri := osDisk.BlobURI; uri != nil {
+			result["blob_uri"] = *uri
 		}
-		if osDisk.DiskSizeGB != nil {
-			result["size_gb"] = *osDisk.DiskSizeGB
+		if diskSizeDB := osDisk.DiskSizeGB; diskSizeDB != nil {
+			result["size_gb"] = *diskSizeDB
 		}
-		if osDisk.ManagedDisk != nil {
-			result["managed_disk_id"] = *osDisk.ManagedDisk.ID
+		if disk := osDisk.ManagedDisk; disk != nil {
+			result["managed_disk_id"] = *disk.ID
 		}
 		result["caching"] = string(osDisk.Caching)
 		result["os_type"] = osDisk.OsType
@@ -310,7 +323,7 @@ func flattenAzureRmImageOSDisk(d *schema.ResourceData, osDisk *compute.ImageOSDi
 	return []interface{}{result}
 }
 
-func flattenAzureRmImageDataDisks(d *schema.ResourceData, diskImages *[]compute.ImageDataDisk) []interface{} {
+func flattenAzureRmImageDataDisks(diskImages *[]compute.ImageDataDisk) []interface{} {
 	result := make([]interface{}, 0)
 
 	if images := diskImages; images != nil {
@@ -336,7 +349,6 @@ func flattenAzureRmImageDataDisks(d *schema.ResourceData, diskImages *[]compute.
 }
 
 func expandAzureRmImageOsDisk(d *schema.ResourceData) (*compute.ImageOSDisk, error) {
-
 	osDisk := &compute.ImageOSDisk{}
 	disks := d.Get("os_disk").(*schema.Set).List()