+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES:
 
 * **New Data Source:** `azurerm_dns_zone` [GH-702]
+* **New Resource:** `azurerm_metric_alertrule` [GH-478]
 
 IMPROVEMENTS: